+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@ by Matt Tropiano et al. (see AUTHORS.txt)
 0.8.0
 -----
 
+- `Fixed` Attempting to start the GUI in a headless environment should error out (Issue #5).
 - `Added` Name suffixes can be added to IWADs added via scan (Enh. #11).
 - `Added` Name suffixes can be added to WADs added via scan.